+++ src/main/java/ru/yourteam/filmorate/controller/CommonFilmsController.java
@@ -1,4 +1,3 @@
-// Контроллер для выдачи списка общих фильмов у пользователей.
 package ru.yourteam.filmorate.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +10,9 @@ import ru.yourteam.filmorate.service.CommonFilmService;
 
 import java.util.List;
 
+/**
+ * Контроллер для выдачи списка общих фильмов у пользователей.
+ */
 @RestController
 @RequestMapping("/films/common")
 public class CommonFilmsController {