+++ src/main/java/ru/yourteam/filmorate/dto/search/MpaDto.java
@@ -1,0 +1,9 @@
+package ru.yourteam.filmorate.dto.search;
+
+import lombok.Data;
+
+@Data
+public class MpaDto {
+    int id;
+    String name;
+}