+++ src/main/java/ru/yourteam/filmorate/dto/searchDto/GenreDto.java
@@ -1,9 +1,0 @@
-package ru.yourteam.filmorate.dto.searchDto;
-
-import lombok.Data;
-
-@Data
-public class GenreDto {
-    int id;
-    String name;
-}